Broke components into their own files
